refactor: address PR review remarks for Aleo add-account flow
