+++ apps/ledger-live-mobile/src/mvvm/features/DeviceSelection/__integrations__/deviceSelection.integration.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { render, screen, withReadOnlyDisabled, renderHook } from "@tests/test-renderer";
+import { render, screen, withReadOnlyDisabled } from "@tests/test-renderer";
 import DeviceSelectionNavigator from "../Navigator";
 import { useRoute, useNavigation } from "@react-navigation/native";
 import { useNavigation as useNavigationCore } from "@react-navigation/core";
@@ -7,11 +7,6 @@ import { discoverDevices } from "@ledgerhq/live-common/hw/index";
 import { DeviceModelId } from "@ledgerhq/types-devices";
 import { of } from "rxjs";
 import { AddAccountContexts } from "../../Accounts/screens/AddAccount/enums";
-import useSelectDeviceViewModel from "../screens/SelectDevice/useSelectDeviceViewModel";
-import { getCustomAddAccountFlow } from "LLM/features/Accounts/customAddAccountFlow";
-import { NavigatorName, ScreenName } from "~/const";
-import type { AppResult } from "@ledgerhq/live-common/hw/actions/app";
-import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 
 const MockUseRoute = useRoute as jest.Mock;
 const mockNavigate = jest.fn();
@@ -43,7 +38,7 @@ jest.mock("@ledgerhq/live-common/hw/index", () => ({
   discoverDevices: jest.fn(),
 }));
 
-jest.mock("LLM/features/Accounts/customAddAccountFlow", () => ({
+jest.mock("LLM/features/Accounts/utils/customAddAccountFlow", () => ({
   getCustomAddAccountFlow: jest.fn(),
 }));
 
@@ -119,108 +114,3 @@ describe("Device Selection feature integration test", () => {
   });
 });
 
-describe("useSelectDeviceViewModel — onResult navigation", () => {
-  const mockGetCustomAddAccountFlow = getCustomAddAccountFlow as jest.Mock;
-  const mockOnDeviceConnected = jest.fn();
-
-  const aleoCurrency: CryptoCurrency = {
-    type: "CryptoCurrency",
-    coinType: 683,
-    scheme: "aleo",
-    id: "aleo",
-    ticker: "ALEO",
-    name: "Aleo",
-    family: "aleo",
-    color: "#000000",
-    managerAppName: "Aleo",
-    units: [],
-    explorerViews: [],
-  };
-
-  const bitcoinCurrency: CryptoCurrency = {
-    type: "CryptoCurrency",
-    coinType: 0,
-    scheme: "bitcoin",
-    id: "bitcoin",
-    ticker: "BTC",
-    name: "Bitcoin",
-    family: "bitcoin",
-    color: "#ffae35",
-    managerAppName: "Bitcoin",
-    units: [],
-    explorerViews: [],
-  };
-
-  const mockMeta = {} as AppResult;
-
-  const getRouteMock = (currency: CryptoCurrency) => {
-    return {
-      params: { context: AddAccountContexts.AddAccounts, currency },
-    } as unknown as Parameters<typeof useSelectDeviceViewModel>[0];
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    (useNavigation as jest.Mock).mockReturnValue({
-      navigate: mockNavigate,
-      addListener: jest.fn(),
-    });
-    (useNavigationCore as jest.Mock).mockReturnValue({
-      navigate: mockNavigate,
-      addListener: jest.fn(),
-    });
-  });
-
-  it("invokes onDeviceConnected and does not navigate to ScanDeviceAccounts when a custom flow is registered", () => {
-    mockGetCustomAddAccountFlow.mockReturnValue({ onDeviceConnected: mockOnDeviceConnected });
-
-    const route = getRouteMock(aleoCurrency);
-
-    const { result } = renderHook(() => useSelectDeviceViewModel(route), {
-      overrideInitialState: withReadOnlyDisabled,
-    });
-
-    result.current.onResult(mockMeta);
-
-    expect(mockOnDeviceConnected).toHaveBeenCalledTimes(1);
-    expect(mockNavigate).not.toHaveBeenCalledWith(
-      NavigatorName.AddAccounts,
-      expect.objectContaining({ screen: ScreenName.ScanDeviceAccounts }),
-    );
-  });
-
-  it("navigates to ScanDeviceAccounts when no custom flow is registered for the currency", () => {
-    mockGetCustomAddAccountFlow.mockReturnValue(null);
-
-    const route = getRouteMock(bitcoinCurrency);
-
-    const { result } = renderHook(() => useSelectDeviceViewModel(route), {
-      overrideInitialState: withReadOnlyDisabled,
-    });
-
-    result.current.onResult(mockMeta);
-
-    expect(mockNavigate).toHaveBeenCalledWith(NavigatorName.AddAccounts, {
-      screen: ScreenName.ScanDeviceAccounts,
-      params: expect.objectContaining({ sourceScreenName: ScreenName.SelectDevice }),
-    });
-    expect(mockOnDeviceConnected).not.toHaveBeenCalled();
-  });
-
-  it("navigates to ScanDeviceAccounts when a custom flow exists but has no onDeviceConnected handler", () => {
-    mockGetCustomAddAccountFlow.mockReturnValue({});
-
-    const route = getRouteMock(bitcoinCurrency);
-
-    const { result } = renderHook(() => useSelectDeviceViewModel(route), {
-      overrideInitialState: withReadOnlyDisabled,
-    });
-
-    result.current.onResult(mockMeta);
-
-    expect(mockNavigate).toHaveBeenCalledWith(NavigatorName.AddAccounts, {
-      screen: ScreenName.ScanDeviceAccounts,
-      params: expect.objectContaining({ sourceScreenName: ScreenName.SelectDevice }),
-    });
-  });
-});
